fix(example): move test harness into integration_test/support
Web integration tests compile via Flutter's web driver, which can't
follow ../ imports out of integration_test/ into a sibling test/ dir —
so the smoke test's '../test/robots/...' imports threw File not found on
web only (macOS compiled them fine). That's the local/CI gap.

Move the shared harness + robots into integration_test/support/ so the
on-device smoke test imports them with no ../ crossing (web-safe). The
host-VM journeys in test/journeys/ reach into ../../integration_test/
support/ — fine, host-VM flutter test has no web-driver ../ restriction.

Verified locally on the real web platform (make test-example-web-jspi:
62/62 incl. UI tests), the matrix (12/12), and make analyze.
